requirements.txt
python-bitcoinrpc==1.0
tornado==5.1.1
